Improving test structure
