+++ package.json
@@ -60,7 +60,8 @@
     "test-debug": "DEBUG='ldnode:*' ./node_modules/mocha/bin/mocha ./test/*.js",
     "test-acl": "./node_modules/mocha/bin/mocha ./test/acl.js",
     "test-params": "./node_modules/mocha/bin/mocha ./test/params.js",
-    "test-http": "./node_modules/mocha/bin/mocha ./test/test.js"
+    "test-http": "./node_modules/mocha/bin/mocha ./test/http.js",
+    "test-formats": "./node_modules/mocha/bin/mocha ./test/http.js"
   },
   "bin": {
     "ldnode": "./bin/ldnode.js"
