+++ test/http.js
@@ -5,7 +5,7 @@ var ldnode = require('../index');
 var fs = require('fs');
 var S = require('string');
 
-describe('ldnode', function() {
+describe('HTTP APIs', function() {
   var address = 'http://localhost:3457';
   var ldp = ldnode.createServer({
     root: __dirname + '/resources',
@@ -14,16 +14,7 @@ describe('ldnode', function() {
 
   var server = supertest(address);
 
-  describe('Hello World', function() {
-      it('Should return "Hello, World!"', function(done) {
-          server.get('/hello.html')
-              .expect('Content-type', /text\/html/)
-              .expect(/Hello, world!/)
-              .expect(200, done);
-      });
-  });
-
-  describe('Root container', function() {
+  describe('GET Root container', function() {
       it('Should exists', function(done) {
           server.get('/')
               .expect(200, done);
@@ -31,42 +22,6 @@ describe('ldnode', function() {
       it('Should be a turtle file by default', function(done) {
           server.get('/')
               .expect('content-type', /text\/turtle/)
-              .expect(200, done);
-      });
-  });
-
-  describe('JSON-LD support', function() {
-      var isValidJSON = function(res) {
-          var json = JSON.parse(res.text);
-      };
-      it('Should return JSON-LD document', function(done) {
-          server.get('/patch-5-initial.ttl')
-              .set('accept', 'application/json+ld')
-              .expect('content-type', /application\/json\+ld/)
-              .expect(200, done);
-      });
-      it('Should return valid JSON', function(done) {
-          server.get('/patch-5-initial.ttl')
-              .set('accept', 'application/json+ld')
-              .expect(isValidJSON)
-              .expect(200, done);
-      });
-  });
-
-  describe('N-Quads support', function() {
-      it('Should return N-Quads document', function(done) {
-          server.get('/patch-5-initial.ttl')
-              .set('accept', 'application/n-quads')
-              .expect('content-type', /application\/n-quads/)
-              .expect(200, done);
-      });
-  });
-
-  describe('n3 support', function() {
-      it('Should return turtle document if content-type set to n3', function(done) {
-          server.get('/patch-5-initial.ttl')
-              .set('accept', 'text/n3')
-              .expect('content-type', /text\/n3/)
               .expect(200, done);
       });
   });
